Interesction accepts
